+++ CHANGELOG.md
@@ -5,6 +5,8 @@ All notable changes to this project will be documented in this file.
 ## [0.10.2] - 2025-09-09
 ### Added
 - Track the number of times poketerm is run and remind heavy users to donate.
+### Changed
+- Limit donation reminders to three colorful messages for users exceeding 10,000 runs.
 
 ## [0.10.1] - 2025-09-09
 ### Added